Generating Menu, Loading selected product, View Count
1) Added partial view for generating menu from database
2) OnClick of menu generating particular category products
3) Saving product view count

+++ ECoupoun.Entities/ProductsJSON.cs
@@ -22,7 +22,7 @@ namespace ECoupoun.Entities
 
     public class Products
     {
-        public int Sku { get; set; }
+        public long Sku { get; set; }
         public string Name { get; set; }
         public string ThumbnailImage { get; set; }
         public string ModelNumber { get; set; }
